requirements.txt
PyMySQL==1.0.2
python-decouple==3.4
python-dotenv==0.19.0
